chore(refactor): Router: Splitting things up into smaller files
