+++ packages/router/src/index.ts
@@ -3,8 +3,10 @@
 // latter of which has closely inspired some of this code).
 
 export { navigate, back } from './history'
-export { Link, NavLink, Redirect } from './links'
+export { NavLink } from './navLink'
+export { Link } from './link'
 export { useLocation, LocationProvider } from './location'
+export { Redirect } from './redirect'
 export {
   usePageLoadingContext,
   PageLoadingContextProvider,
