+++ packages/router/src/navLink.tsx
@@ -1,58 +1,18 @@
-import { forwardRef, useEffect } from 'react'
+import { forwardRef } from 'react'
 
-import type { NavigateOptions } from './history'
 import { navigate } from './history'
+import type { LinkProps } from './link'
 import { useMatch } from './useMatch'
 import type { FlattenSearchParams } from './util'
 import { flattenSearchParams } from './util'
 
-interface LinkProps {
-  to: string
-  onClick?: React.MouseEventHandler<HTMLAnchorElement>
-}
-
-const Link = forwardRef<
-  HTMLAnchorElement,
-  LinkProps & React.AnchorHTMLAttributes<HTMLAnchorElement>
->(({ to, onClick, ...rest }, ref) => (
-  <a
-    href={to}
-    ref={ref}
-    {...rest}
-    onClick={(event) => {
-      if (
-        event.button !== 0 ||
-        event.altKey ||
-        event.ctrlKey ||
-        event.metaKey ||
-        event.shiftKey
-      ) {
-        return
-      }
-
-      event.preventDefault()
-
-      if (onClick) {
-        const result = onClick(event)
-        if (typeof result !== 'boolean' || result) {
-          navigate(to)
-        }
-      } else {
-        navigate(to)
-      }
-    }}
-  />
-))
-
-interface NavLinkProps {
-  to: string
+interface NavLinkProps extends LinkProps {
   activeClassName: string
   activeMatchParams?: FlattenSearchParams
   matchSubPaths?: boolean
-  onClick?: React.MouseEventHandler<HTMLAnchorElement>
 }
 
-const NavLink = forwardRef<
+export const NavLink = forwardRef<
   HTMLAnchorElement,
   NavLinkProps & React.AnchorHTMLAttributes<HTMLAnchorElement>
 >(
@@ -111,22 +71,3 @@ const NavLink = forwardRef<
     )
   }
 )
-
-interface RedirectProps {
-  /** The path to redirect to */
-  to: string
-  options?: NavigateOptions
-}
-
-/**
- * A declarative way to redirect to a route name
- */
-const Redirect = ({ to, options }: RedirectProps) => {
-  useEffect(() => {
-    navigate(to, options)
-  }, [to, options])
-
-  return null
-}
-
-export { Link, NavLink, Redirect }
