+++ packages/router/src/params.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react'
 
-import { createNamedContext } from './util'
+import { createNamedContext } from './createNamedContext'
 
 export interface ParamsContextProps {
   params: Record<string, string>
